[PM-2966] Move MockConnectionProvider to separate file
